requirements.txt
Django
mysqlclient
google-generativeai
python-dotenv
Pillow